+++ program-runtime/Cargo.toml
@@ -10,6 +10,7 @@ documentation = "https://docs.rs/solana-program-runtime"
 edition = "2018"
 
 [dependencies]
+bincode = "1.3.3"
 libc = "0.2.101"
 libloading = "0.7.0"
 log = "0.4.14"
