+++ runtime/src/lib.rs
@@ -32,9 +32,7 @@ pub mod genesis_utils;
 pub mod hardened_unpack;
 pub mod in_mem_accounts_index;
 pub mod inline_spl_token_v2_0;
-pub mod instruction_recorder;
 pub mod loader_utils;
-pub mod log_collector;
 pub mod message_processor;
 pub mod neon_evm_program;
 pub mod non_circulating_supply;
