+++ runtime/src/system_instruction_processor.rs
@@ -481,25 +481,24 @@ mod tests {
     use super::*;
     use crate::{bank::Bank, bank_client::BankClient};
     use bincode::serialize;
+    use solana_program_runtime::invoke_context::{mock_process_instruction, ThisInvokeContext};
     #[allow(deprecated)]
-    use solana_sdk::sysvar::recent_blockhashes::IterItem;
     use solana_sdk::{
         account::{self, Account, AccountSharedData},
         client::SyncClient,
+        feature_set::FeatureSet,
         genesis_config::create_genesis_config,
         hash::{hash, Hash},
         instruction::{AccountMeta, Instruction, InstructionError},
-        keyed_account::create_keyed_accounts_unified,
         message::Message,
-        nonce, nonce_account,
-        process_instruction::MockInvokeContext,
-        recent_blockhashes_account,
+        nonce, nonce_account, recent_blockhashes_account,
         signature::{Keypair, Signer},
         system_instruction, system_program, sysvar,
+        sysvar::recent_blockhashes::IterItem,
         transaction::TransactionError,
     };
-    use std::cell::RefCell;
     use std::sync::Arc;
+    use std::{cell::RefCell, rc::Rc};
 
     impl From<Pubkey> for Address {
         fn from(address: Pubkey) -> Self {
@@ -511,50 +510,46 @@ mod tests {
     }
 
     fn process_instruction(
-        owner: &Pubkey,
         instruction_data: &[u8],
-        keyed_accounts: &[(bool, bool, &Pubkey, &RefCell<AccountSharedData>)],
+        keyed_accounts: &[(bool, bool, Pubkey, Rc<RefCell<AccountSharedData>>)],
     ) -> Result<(), InstructionError> {
-        let processor_account = RefCell::new(AccountSharedData::from(Account {
-            owner: solana_sdk::native_loader::id(),
-            ..Account::default()
-        }));
-        let mut keyed_accounts = keyed_accounts.to_vec();
-        keyed_accounts.insert(0, (false, false, owner, &processor_account));
-        super::process_instruction(
-            1,
+        mock_process_instruction(
+            &system_program::id(),
+            Vec::new(),
             instruction_data,
-            &mut MockInvokeContext::new(owner, create_keyed_accounts_unified(&keyed_accounts)),
+            keyed_accounts,
+            super::process_instruction,
         )
     }
 
-    fn create_default_account() -> RefCell<AccountSharedData> {
-        RefCell::new(AccountSharedData::default())
+    fn create_default_account() -> Rc<RefCell<AccountSharedData>> {
+        AccountSharedData::new_ref(0, 0, &Pubkey::new_unique())
     }
-    fn create_default_recent_blockhashes_account() -> RefCell<AccountSharedData> {
-        RefCell::new(
+    fn create_default_recent_blockhashes_account() -> Rc<RefCell<AccountSharedData>> {
+        Rc::new(RefCell::new(
             #[allow(deprecated)]
             recent_blockhashes_account::create_account_with_data_for_test(
                 vec![IterItem(0u64, &Hash::default(), 0); sysvar::recent_blockhashes::MAX_ENTRIES]
                     .into_iter(),
             ),
-        )
+        ))
     }
-    fn create_default_rent_account() -> RefCell<AccountSharedData> {
-        RefCell::new(account::create_account_shared_data_for_test(&Rent::free()))
+    fn create_default_rent_account() -> Rc<RefCell<AccountSharedData>> {
+        Rc::new(RefCell::new(account::create_account_shared_data_for_test(
+            &Rent::free(),
+        )))
     }
 
     #[test]
     fn test_create_account() {
         let new_owner = Pubkey::new(&[9; 32]);
-        let from = solana_sdk::pubkey::new_rand();
-        let to = solana_sdk::pubkey::new_rand();
+        let from = Pubkey::new_unique();
+        let to = Pubkey::new_unique();
         let from_account = AccountSharedData::new_ref(100, 0, &system_program::id());
         let to_account = AccountSharedData::new_ref(0, 0, &Pubkey::default());
 
         assert_eq!(
             process_instruction(
-                &Pubkey::default(),
                 &bincode::serialize(&SystemInstruction::CreateAccount {
                     lamports: 50,
                     space: 2,
@@ -562,8 +557,8 @@ mod tests {
                 })
                 .unwrap(),
                 &[
-                    (true, false, &from, &from_account),
-                    (true, false, &to, &to_account),
+                    (true, false, from, from_account.clone()),
+                    (true, false, to, to_account.clone()),
                 ],
             ),
             Ok(())
@@ -577,7 +572,7 @@ mod tests {
     #[test]
     fn test_create_account_with_seed() {
         let new_owner = Pubkey::new(&[9; 32]);
-        let from = solana_sdk::pubkey::new_rand();
+        let from = Pubkey::new_unique();
         let seed = "shiny pepper";
         let to = Pubkey::create_with_seed(&from, seed, &new_owner).unwrap();
         let from_account = AccountSharedData::new_ref(100, 0, &system_program::id());
@@ -585,7 +580,6 @@ mod tests {
 
         assert_eq!(
             process_instruction(
-                &Pubkey::default(),
                 &bincode::serialize(&SystemInstruction::CreateAccountWithSeed {
                     base: from,
                     seed: seed.to_string(),
@@ -595,8 +589,8 @@ mod tests {
                 })
                 .unwrap(),
                 &[
-                    (true, false, &from, &from_account),
-                    (false, false, &to, &to_account),
+                    (true, false, from, from_account.clone()),
+                    (false, false, to, to_account.clone()),
                 ],
             ),
             Ok(())
@@ -610,8 +604,8 @@ mod tests {
     #[test]
     fn test_create_account_with_seed_separate_base_account() {
         let new_owner = Pubkey::new(&[9; 32]);
-        let from = solana_sdk::pubkey::new_rand();
-        let base = solana_sdk::pubkey::new_rand();
+        let from = Pubkey::new_unique();
+        let base = Pubkey::new_unique();
         let seed = "shiny pepper";
         let to = Pubkey::create_with_seed(&base, seed, &new_owner).unwrap();
         let from_account = AccountSharedData::new_ref(100, 0, &system_program::id());
@@ -620,7 +614,6 @@ mod tests {
 
         assert_eq!(
             process_instruction(
-                &Pubkey::default(),
                 &bincode::serialize(&SystemInstruction::CreateAccountWithSeed {
                     base,
                     seed: seed.to_string(),
@@ -630,9 +623,9 @@ mod tests {
                 })
                 .unwrap(),
                 &[
-                    (true, false, &from, &from_account),
-                    (false, false, &to, &to_account),
-                    (true, false, &base, &base_account),
+                    (true, false, from, from_account.clone()),
+                    (false, false, to, to_account.clone()),
+                    (true, false, base, base_account),
                 ],
             ),
             Ok(())
@@ -645,36 +638,30 @@ mod tests {
 
     #[test]
     fn test_address_create_with_seed_mismatch() {
-        let from = solana_sdk::pubkey::new_rand();
+        let invoke_context = ThisInvokeContext::new_mock(&[], &[]);
+        let from = Pubkey::new_unique();
         let seed = "dull boy";
-        let to = solana_sdk::pubkey::new_rand();
-        let owner = solana_sdk::pubkey::new_rand();
+        let to = Pubkey::new_unique();
+        let owner = Pubkey::new_unique();
 
         assert_eq!(
-            Address::create(
-                &to,
-                Some((&from, seed, &owner)),
-                &MockInvokeContext::new(&Pubkey::default(), vec![]),
-            ),
+            Address::create(&to, Some((&from, seed, &owner)), &invoke_context),
             Err(SystemError::AddressWithSeedMismatch.into())
         );
     }
 
     #[test]
     fn test_create_account_with_seed_missing_sig() {
+        let invoke_context = ThisInvokeContext::new_mock(&[], &[]);
         let new_owner = Pubkey::new(&[9; 32]);
-        let from = solana_sdk::pubkey::new_rand();
+        let from = Pubkey::new_unique();
         let seed = "dull boy";
         let to = Pubkey::create_with_seed(&from, seed, &new_owner).unwrap();
 
         let from_account = AccountSharedData::new_ref(100, 0, &system_program::id());
         let to_account = AccountSharedData::new_ref(0, 0, &Pubkey::default());
-        let to_address = Address::create(
-            &to,
-            Some((&from, seed, &new_owner)),
-            &MockInvokeContext::new(&Pubkey::default(), vec![]),
-        )
-        .unwrap();
+        let to_address =
+            Address::create(&to, Some((&from, seed, &new_owner)), &invoke_context).unwrap();
 
         assert_eq!(
             create_account(
@@ -685,7 +672,7 @@ mod tests {
                 2,
                 &new_owner,
                 &HashSet::new(),
-                &MockInvokeContext::new(&Pubkey::default(), vec![]),
+                &invoke_context,
             ),
             Err(InstructionError::MissingRequiredSignature)
         );
@@ -695,6 +682,7 @@ mod tests {
 
     #[test]
     fn test_create_with_zero_lamports() {
+        let invoke_context = ThisInvokeContext::new_mock(&[], &[]);
         // create account with zero lamports transferred
         let new_owner = Pubkey::new(&[9; 32]);
         let from = Pubkey::new_unique();
@@ -712,7 +700,7 @@ mod tests {
                 2,
                 &new_owner,
                 &[from, to].iter().cloned().collect::<HashSet<_>>(),
-                &MockInvokeContext::new(&Pubkey::default(), vec![]),
+                &invoke_context,
             ),
             Ok(())
         );
@@ -728,12 +716,13 @@ mod tests {
 
     #[test]
     fn test_create_negative_lamports() {
+        let invoke_context = ThisInvokeContext::new_mock(&[], &[]);
         // Attempt to create account with more lamports than remaining in from_account
         let new_owner = Pubkey::new(&[9; 32]);
-        let from = solana_sdk::pubkey::new_rand();
+        let from = Pubkey::new_unique();
         let from_account = AccountSharedData::new_ref(100, 0, &system_program::id());
 
-        let to = solana_sdk::pubkey::new_rand();
+        let to = Pubkey::new_unique();
         let to_account = AccountSharedData::new_ref(0, 0, &Pubkey::default());
 
         let result = create_account(
@@ -744,17 +733,18 @@ mod tests {
             2,
             &new_owner,
             &[from, to].iter().cloned().collect::<HashSet<_>>(),
-            &MockInvokeContext::new(&Pubkey::default(), vec![]),
+            &invoke_context,
         );
         assert_eq!(result, Err(SystemError::ResultWithNegativeLamports.into()));
     }
 
     #[test]
     fn test_request_more_than_allowed_data_length() {
+        let invoke_context = ThisInvokeContext::new_mock(&[], &[]);
         let from_account = AccountSharedData::new_ref(100, 0, &system_program::id());
-        let from = solana_sdk::pubkey::new_rand();
+        let from = Pubkey::new_unique();
         let to_account = AccountSharedData::new_ref(0, 0, &system_program::id());
-        let to = solana_sdk::pubkey::new_rand();
+        let to = Pubkey::new_unique();
 
         let signers = &[from, to].iter().cloned().collect::<HashSet<_>>();
         let address = &to.into();
@@ -768,7 +758,7 @@ mod tests {
             MAX_PERMITTED_DATA_LENGTH + 1,
             &system_program::id(),
             signers,
-            &MockInvokeContext::new(&Pubkey::default(), vec![]),
+            &invoke_context,
         );
         assert!(result.is_err());
         assert_eq!(
@@ -785,7 +775,7 @@ mod tests {
             MAX_PERMITTED_DATA_LENGTH,
             &system_program::id(),
             signers,
-            &MockInvokeContext::new(&Pubkey::default(), vec![]),
+            &invoke_context,
         );
         assert!(result.is_ok());
         assert_eq!(to_account.borrow().lamports(), 50);
@@ -797,13 +787,14 @@ mod tests {
 
     #[test]
     fn test_create_already_in_use() {
+        let invoke_context = ThisInvokeContext::new_mock(&[], &[]);
         // Attempt to create system account in account already owned by another program
         let new_owner = Pubkey::new(&[9; 32]);
-        let from = solana_sdk::pubkey::new_rand();
+        let from = Pubkey::new_unique();
         let from_account = AccountSharedData::new_ref(100, 0, &system_program::id());
 
         let original_program_owner = Pubkey::new(&[5; 32]);
-        let owned_key = solana_sdk::pubkey::new_rand();
+        let owned_key = Pubkey::new_unique();
         let owned_account = AccountSharedData::new_ref(0, 0, &original_program_owner);
         let unchanged_account = owned_account.clone();
 
@@ -818,7 +809,7 @@ mod tests {
             2,
             &new_owner,
             signers,
-            &MockInvokeContext::new(&Pubkey::default(), vec![]),
+            &invoke_context,
         );
         assert_eq!(result, Err(SystemError::AccountAlreadyInUse.into()));
 
@@ -837,7 +828,7 @@ mod tests {
             2,
             &new_owner,
             signers,
-            &MockInvokeContext::new(&Pubkey::default(), vec![]),
+            &invoke_context,
         );
         assert_eq!(result, Err(SystemError::AccountAlreadyInUse.into()));
         let from_lamports = from_account.borrow().lamports();
@@ -855,7 +846,7 @@ mod tests {
             2,
             &new_owner,
             signers,
-            &MockInvokeContext::new(&Pubkey::default(), vec![]),
+            &invoke_context,
         );
         assert_eq!(result, Err(SystemError::AccountAlreadyInUse.into()));
         assert_eq!(from_lamports, 100);
@@ -864,12 +855,13 @@ mod tests {
 
     #[test]
     fn test_create_unsigned() {
+        let invoke_context = ThisInvokeContext::new_mock(&[], &[]);
         // Attempt to create an account without signing the transfer
         let new_owner = Pubkey::new(&[9; 32]);
-        let from = solana_sdk::pubkey::new_rand();
+        let from = Pubkey::new_unique();
         let from_account = AccountSharedData::new_ref(100, 0, &system_program::id());
 
-        let owned_key = solana_sdk::pubkey::new_rand();
+        let owned_key = Pubkey::new_unique();
         let owned_account = AccountSharedData::new_ref(0, 0, &Pubkey::default());
 
         let owned_address = owned_key.into();
@@ -883,7 +875,7 @@ mod tests {
             2,
             &new_owner,
             &[owned_key].iter().cloned().collect::<HashSet<_>>(),
-            &MockInvokeContext::new(&Pubkey::default(), vec![]),
+            &invoke_context,
         );
         assert_eq!(result, Err(InstructionError::MissingRequiredSignature));
 
@@ -897,7 +889,7 @@ mod tests {
             2,
             &new_owner,
             &[from].iter().cloned().collect::<HashSet<_>>(),
-            &MockInvokeContext::new(&Pubkey::default(), vec![]),
+            &invoke_context,
         );
         assert_eq!(result, Err(InstructionError::MissingRequiredSignature));
 
@@ -911,18 +903,19 @@ mod tests {
             2,
             &new_owner,
             &[owned_key].iter().cloned().collect::<HashSet<_>>(),
-            &MockInvokeContext::new(&Pubkey::default(), vec![]),
+            &invoke_context,
         );
         assert_eq!(result, Err(InstructionError::MissingRequiredSignature));
     }
 
     #[test]
     fn test_create_sysvar_invalid_id_with_feature() {
+        let invoke_context = ThisInvokeContext::new_mock(&[], &[]);
         // Attempt to create system account in account already owned by another program
-        let from = solana_sdk::pubkey::new_rand();
+        let from = Pubkey::new_unique();
         let from_account = AccountSharedData::new_ref(100, 0, &system_program::id());
 
-        let to = solana_sdk::pubkey::new_rand();
+        let to = Pubkey::new_unique();
         let to_account = AccountSharedData::new_ref(0, 0, &system_program::id());
 
         let signers = [from, to].iter().cloned().collect::<HashSet<_>>();
@@ -937,18 +930,31 @@ mod tests {
             2,
             &sysvar::id(),
             &signers,
-            &MockInvokeContext::new(&Pubkey::default(), vec![]),
+            &invoke_context,
         );
         assert_eq!(result, Ok(()));
     }
 
     #[test]
     fn test_create_sysvar_invalid_id_without_feature() {
+        let mut feature_set = FeatureSet::all_enabled();
+        feature_set
+            .active
+            .remove(&feature_set::rent_for_sysvars::id());
+        feature_set
+            .inactive
+            .insert(feature_set::rent_for_sysvars::id());
+        let invoke_context = ThisInvokeContext::new_mock_with_sysvars_and_features(
+            &[],
+            &[],
+            &[],
+            Arc::new(feature_set),
+        );
         // Attempt to create system account in account already owned by another program
-        let from = solana_sdk::pubkey::new_rand();
+        let from = Pubkey::new_unique();
         let from_account = AccountSharedData::new_ref(100, 0, &system_program::id());
 
-        let to = solana_sdk::pubkey::new_rand();
+        let to = Pubkey::new_unique();
         let to_account = AccountSharedData::new_ref(0, 0, &system_program::id());
 
         let signers = [from, to].iter().cloned().collect::<HashSet<_>>();
@@ -962,24 +968,20 @@ mod tests {
             2,
             &sysvar::id(),
             &signers,
-            &MockInvokeContext {
-                disabled_features: vec![feature_set::rent_for_sysvars::id()]
-                    .into_iter()
-                    .collect(),
-                ..MockInvokeContext::new(&Pubkey::default(), vec![])
-            },
+            &invoke_context,
         );
         assert_eq!(result, Err(SystemError::InvalidProgramId.into()));
     }
 
     #[test]
     fn test_create_data_populated() {
+        let invoke_context = ThisInvokeContext::new_mock(&[], &[]);
         // Attempt to create system account in account with populated data
         let new_owner = Pubkey::new(&[9; 32]);
-        let from = solana_sdk::pubkey::new_rand();
+        let from = Pubkey::new_unique();
         let from_account = AccountSharedData::new_ref(100, 0, &system_program::id());
 
-        let populated_key = solana_sdk::pubkey::new_rand();
+        let populated_key = Pubkey::new_unique();
         let populated_account = AccountSharedData::from(Account {
             data: vec![0, 1, 2, 3],
             ..Account::default()
@@ -1000,14 +1002,15 @@ mod tests {
             2,
             &new_owner,
             &signers,
-            &MockInvokeContext::new(&Pubkey::default(), vec![]),
+            &invoke_context,
         );
         assert_eq!(result, Err(SystemError::AccountAlreadyInUse.into()));
     }
 
     #[test]
     fn test_create_from_account_is_nonce_fail() {
-        let nonce = solana_sdk::pubkey::new_rand();
+        let invoke_context = ThisInvokeContext::new_mock(&[], &[]);
+        let nonce = Pubkey::new_unique();
         let nonce_account = AccountSharedData::new_ref_data(
             42,
             &nonce::state::Versions::new_current(nonce::State::Initialized(
@@ -1017,7 +1020,7 @@ mod tests {
         )
         .unwrap();
         let from = KeyedAccount::new(&nonce, true, &nonce_account);
-        let new = solana_sdk::pubkey::new_rand();
+        let new = Pubkey::new_unique();
 
         let new_account = AccountSharedData::new_ref(0, 0, &system_program::id());
 
@@ -1032,9 +1035,9 @@ mod tests {
                 &new_address,
                 42,
                 0,
-                &solana_sdk::pubkey::new_rand(),
+                &Pubkey::new_unique(),
                 &signers,
-                &MockInvokeContext::new(&Pubkey::default(), vec![]),
+                &invoke_context,
             ),
             Err(InstructionError::InvalidArgument),
         );
@@ -1042,8 +1045,9 @@ mod tests {
 
     #[test]
     fn test_assign() {
+        let invoke_context = ThisInvokeContext::new_mock(&[], &[]);
         let new_owner = Pubkey::new(&[9; 32]);
-        let pubkey = solana_sdk::pubkey::new_rand();
+        let pubkey = Pubkey::new_unique();
         let mut account = AccountSharedData::new(100, 0, &system_program::id());
 
         assert_eq!(
@@ -1052,7 +1056,7 @@ mod tests {
                 &pubkey.into(),
                 &new_owner,
                 &HashSet::new(),
-                &MockInvokeContext::new(&Pubkey::default(), vec![]),
+                &invoke_context,
             ),
             Err(InstructionError::MissingRequiredSignature)
         );
@@ -1064,17 +1068,16 @@ mod tests {
                 &pubkey.into(),
                 &system_program::id(),
                 &HashSet::new(),
-                &MockInvokeContext::new(&Pubkey::default(), vec![]),
+                &invoke_context,
             ),
             Ok(())
         );
 
-        let account = RefCell::new(account);
+        let account = Rc::new(RefCell::new(account));
         assert_eq!(
             process_instruction(
-                &Pubkey::default(),
                 &bincode::serialize(&SystemInstruction::Assign { owner: new_owner }).unwrap(),
-                &[(true, false, &pubkey, &account)],
+                &[(true, false, pubkey, account)],
             ),
             Ok(())
         );
@@ -1082,9 +1085,9 @@ mod tests {
 
     #[test]
     fn test_assign_to_sysvar_with_feature() {
+        let invoke_context = ThisInvokeContext::new_mock(&[], &[]);
         let new_owner = sysvar::id();
-
-        let from = solana_sdk::pubkey::new_rand();
+        let from = Pubkey::new_unique();
         let mut from_account = AccountSharedData::new(100, 0, &system_program::id());
 
         assert_eq!(
@@ -1093,7 +1096,7 @@ mod tests {
                 &from.into(),
                 &new_owner,
                 &[from].iter().cloned().collect::<HashSet<_>>(),
-                &MockInvokeContext::new(&Pubkey::default(), vec![]),
+                &invoke_context,
             ),
             Ok(())
         );
@@ -1101,9 +1104,21 @@ mod tests {
 
     #[test]
     fn test_assign_to_sysvar_without_feature() {
+        let mut feature_set = FeatureSet::all_enabled();
+        feature_set
+            .active
+            .remove(&feature_set::rent_for_sysvars::id());
+        feature_set
+            .inactive
+            .insert(feature_set::rent_for_sysvars::id());
+        let invoke_context = ThisInvokeContext::new_mock_with_sysvars_and_features(
+            &[],
+            &[],
+            &[],
+            Arc::new(feature_set),
+        );
         let new_owner = sysvar::id();
-
-        let from = solana_sdk::pubkey::new_rand();
+        let from = Pubkey::new_unique();
         let mut from_account = AccountSharedData::new(100, 0, &system_program::id());
 
         assert_eq!(
@@ -1112,12 +1127,7 @@ mod tests {
                 &from.into(),
                 &new_owner,
                 &[from].iter().cloned().collect::<HashSet<_>>(),
-                &MockInvokeContext {
-                    disabled_features: vec![feature_set::rent_for_sysvars::id()]
-                        .into_iter()
-                        .collect(),
-                    ..MockInvokeContext::new(&Pubkey::default(), vec![])
-                },
+                &invoke_context,
             ),
             Err(SystemError::InvalidProgramId.into())
         );
@@ -1127,40 +1137,31 @@ mod tests {
     fn test_process_bogus_instruction() {
         // Attempt to assign with no accounts
         let instruction = SystemInstruction::Assign {
-            owner: solana_sdk::pubkey::new_rand(),
+            owner: Pubkey::new_unique(),
         };
         let data = serialize(&instruction).unwrap();
-        let result = process_instruction(&system_program::id(), &data, &[]);
+        let result = process_instruction(&data, &[]);
         assert_eq!(result, Err(InstructionError::NotEnoughAccountKeys));
 
         // Attempt to transfer with no destination
-        let from = solana_sdk::pubkey::new_rand();
+        let from = Pubkey::new_unique();
         let from_account = AccountSharedData::new_ref(100, 0, &system_program::id());
         let instruction = SystemInstruction::Transfer { lamports: 0 };
         let data = serialize(&instruction).unwrap();
-        let result = process_instruction(
-            &system_program::id(),
-            &data,
-            &[(true, false, &from, &from_account)],
-        );
+        let result = process_instruction(&data, &[(true, false, from, from_account)]);
         assert_eq!(result, Err(InstructionError::NotEnoughAccountKeys));
     }
 
     #[test]
     fn test_transfer_lamports() {
-        let from = solana_sdk::pubkey::new_rand();
+        let invoke_context = ThisInvokeContext::new_mock(&[], &[]);
+        let from = Pubkey::new_unique();
         let from_account = AccountSharedData::new_ref(100, 0, &Pubkey::new(&[2; 32])); // account owner should not matter
         let to = Pubkey::new(&[3; 32]);
         let to_account = AccountSharedData::new_ref(1, 0, &to); // account owner should not matter
         let from_keyed_account = KeyedAccount::new(&from, true, &from_account);
         let to_keyed_account = KeyedAccount::new(&to, false, &to_account);
-        transfer(
-            &from_keyed_account,
-            &to_keyed_account,
-            50,
-            &MockInvokeContext::new(&Pubkey::default(), vec![]),
-        )
-        .unwrap();
+        transfer(&from_keyed_account, &to_keyed_account, 50, &invoke_context).unwrap();
         let from_lamports = from_keyed_account.account.borrow().lamports();
         let to_lamports = to_keyed_account.account.borrow().lamports();
         assert_eq!(from_lamports, 50);
@@ -1168,24 +1169,13 @@ mod tests {
 
         // Attempt to move more lamports than remaining in from_account
         let from_keyed_account = KeyedAccount::new(&from, true, &from_account);
-        let result = transfer(
-            &from_keyed_account,
-            &to_keyed_account,
-            100,
-            &MockInvokeContext::new(&Pubkey::default(), vec![]),
-        );
+        let result = transfer(&from_keyed_account, &to_keyed_account, 100, &invoke_context);
         assert_eq!(result, Err(SystemError::ResultWithNegativeLamports.into()));
         assert_eq!(from_keyed_account.account.borrow().lamports(), 50);
         assert_eq!(to_keyed_account.account.borrow().lamports(), 51);
 
         // test signed transfer of zero
-        assert!(transfer(
-            &from_keyed_account,
-            &to_keyed_account,
-            0,
-            &MockInvokeContext::new(&Pubkey::default(), vec![]),
-        )
-        .is_ok());
+        assert!(transfer(&from_keyed_account, &to_keyed_account, 0, &invoke_context).is_ok());
         assert_eq!(from_keyed_account.account.borrow().lamports(), 50);
         assert_eq!(to_keyed_account.account.borrow().lamports(), 51);
 
@@ -1193,12 +1183,7 @@ mod tests {
         let from_keyed_account = KeyedAccount::new(&from, false, &from_account);
 
         assert_eq!(
-            transfer(
-                &from_keyed_account,
-                &to_keyed_account,
-                0,
-                &MockInvokeContext::new(&Pubkey::default(), vec![]),
-            ),
+            transfer(&from_keyed_account, &to_keyed_account, 0, &invoke_context),
             Err(InstructionError::MissingRequiredSignature)
         );
         assert_eq!(from_keyed_account.account.borrow().lamports(), 50);
@@ -1207,7 +1192,8 @@ mod tests {
 
     #[test]
     fn test_transfer_with_seed() {
-        let base = solana_sdk::pubkey::new_rand();
+        let invoke_context = ThisInvokeContext::new_mock(&[], &[]);
+        let base = Pubkey::new_unique();
         let base_account = AccountSharedData::new_ref(100, 0, &Pubkey::new(&[2; 32])); // account owner should not matter
         let from_base_keyed_account = KeyedAccount::new(&base, true, &base_account);
         let from_seed = "42";
@@ -1225,7 +1211,7 @@ mod tests {
             &from_owner,
             &to_keyed_account,
             50,
-            &MockInvokeContext::new(&Pubkey::default(), vec![]),
+            &invoke_context,
         )
         .unwrap();
         let from_lamports = from_keyed_account.account.borrow().lamports();
@@ -1242,7 +1228,7 @@ mod tests {
             &from_owner,
             &to_keyed_account,
             100,
-            &MockInvokeContext::new(&Pubkey::default(), vec![]),
+            &invoke_context,
         );
         assert_eq!(result, Err(SystemError::ResultWithNegativeLamports.into()));
         assert_eq!(from_keyed_account.account.borrow().lamports(), 50);
@@ -1257,7 +1243,7 @@ mod tests {
             &from_owner,
             &to_keyed_account,
             0,
-            &MockInvokeContext::new(&Pubkey::default(), vec![]),
+            &invoke_context,
         )
         .is_ok());
         assert_eq!(from_keyed_account.account.borrow().lamports(), 50);
@@ -1266,7 +1252,8 @@ mod tests {
 
     #[test]
     fn test_transfer_lamports_from_nonce_account_fail() {
-        let from = solana_sdk::pubkey::new_rand();
+        let invoke_context = ThisInvokeContext::new_mock(&[], &[]);
+        let from = Pubkey::new_unique();
         let from_account = AccountSharedData::new_ref_data(
             100,
             &nonce::state::Versions::new_current(nonce::State::Initialized(nonce::state::Data {
@@ -1288,7 +1275,7 @@ mod tests {
                 &KeyedAccount::new(&from, true, &from_account),
                 &KeyedAccount::new(&to, false, &to_account),
                 50,
-                &MockInvokeContext::new(&Pubkey::default(), vec![]),
+                &invoke_context,
             ),
             Err(InstructionError::InvalidArgument),
         )
@@ -1303,7 +1290,7 @@ mod tests {
         let alice_keypair = Keypair::new();
         let alice_pubkey = alice_keypair.pubkey();
         let seed = "seed";
-        let owner = solana_sdk::pubkey::new_rand();
+        let owner = Pubkey::new_unique();
         let alice_with_seed = Pubkey::create_with_seed(&alice_pubkey, seed, &owner).unwrap();
 
         bank_client
@@ -1344,8 +1331,8 @@ mod tests {
         let alice_pubkey = alice_keypair.pubkey();
         let bob_pubkey = bob_keypair.pubkey();
 
-        let program = solana_sdk::pubkey::new_rand();
-        let collector = solana_sdk::pubkey::new_rand();
+        let program = Pubkey::new_unique();
+        let collector = Pubkey::new_unique();
 
         let mint_lamports = 10000;
         let len1 = 123;
@@ -1371,7 +1358,7 @@ mod tests {
         let bank = Arc::new(Bank::new_from_parent(&bank, &collector, bank.slot() + 1));
         let bank_client = BankClient::new_shared(&bank);
         bank_client
-            .transfer_and_confirm(50, &alice_keypair, &solana_sdk::pubkey::new_rand())
+            .transfer_and_confirm(50, &alice_keypair, &Pubkey::new_unique())
             .unwrap();
 
         // super fun time; callback chooses to .clean_accounts(None) or not
@@ -1415,7 +1402,7 @@ mod tests {
         let alice_keypair = Keypair::new();
         let alice_pubkey = alice_keypair.pubkey();
         let seed = "seed";
-        let owner = solana_sdk::pubkey::new_rand();
+        let owner = Pubkey::new_unique();
         let alice_with_seed = Pubkey::create_with_seed(&alice_pubkey, seed, &owner).unwrap();
 
         bank_client
@@ -1474,30 +1461,25 @@ mod tests {
     }
 
     fn process_nonce_instruction(instruction: &Instruction) -> Result<(), InstructionError> {
-        let accounts: Vec<_> = instruction
+        let accounts = instruction.accounts.iter().map(|meta| {
+            #[allow(deprecated)]
+            if sysvar::recent_blockhashes::check_id(&meta.pubkey) {
+                create_default_recent_blockhashes_account()
+            } else if sysvar::rent::check_id(&meta.pubkey) {
+                Rc::new(RefCell::new(account::create_account_shared_data_for_test(
+                    &Rent::free(),
+                )))
+            } else {
+                AccountSharedData::new_ref(0, 0, &Pubkey::new_unique())
+            }
+        });
+        let keyed_accounts: Vec<_> = instruction
             .accounts
             .iter()
-            .map(|meta| {
-                #[allow(deprecated)]
-                RefCell::new(if sysvar::recent_blockhashes::check_id(&meta.pubkey) {
-                    create_default_recent_blockhashes_account().into_inner()
-                } else if sysvar::rent::check_id(&meta.pubkey) {
-                    account::create_account_shared_data_for_test(&Rent::free())
-                } else {
-                    AccountSharedData::default()
-                })
-            })
+            .zip(accounts)
+            .map(|(meta, account)| (meta.is_signer, meta.is_writable, meta.pubkey, account))
             .collect();
-
-        {
-            let keyed_accounts: Vec<_> = instruction
-                .accounts
-                .iter()
-                .zip(accounts.iter())
-                .map(|(meta, account)| (meta.is_signer, false, &meta.pubkey, account))
-                .collect();
-            process_instruction(&Pubkey::default(), &instruction.data, &keyed_accounts)
-        }
+        process_instruction(&instruction.data, &keyed_accounts)
     }
 
     #[test]
@@ -1515,7 +1497,6 @@ mod tests {
     fn test_process_nonce_ix_no_keyed_accs_fail() {
         assert_eq!(
             process_instruction(
-                &Pubkey::default(),
                 &serialize(&SystemInstruction::AdvanceNonceAccount).unwrap(),
                 &[],
             ),
@@ -1527,9 +1508,8 @@ mod tests {
     fn test_process_nonce_ix_only_nonce_acc_fail() {
         assert_eq!(
             process_instruction(
-                &Pubkey::default(),
                 &serialize(&SystemInstruction::AdvanceNonceAccount).unwrap(),
-                &[(true, false, &Pubkey::default(), &create_default_account())],
+                &[(true, false, Pubkey::default(), create_default_account())],
             ),
             Err(InstructionError::NotEnoughAccountKeys),
         );
@@ -1539,16 +1519,15 @@ mod tests {
     fn test_process_nonce_ix_bad_recent_blockhash_state_fail() {
         assert_eq!(
             process_instruction(
-                &Pubkey::default(),
                 &serialize(&SystemInstruction::AdvanceNonceAccount).unwrap(),
                 &[
-                    (true, false, &Pubkey::default(), &create_default_account()),
+                    (true, false, Pubkey::default(), create_default_account()),
                     (
                         false,
                         false,
                         #[allow(deprecated)]
-                        &sysvar::recent_blockhashes::id(),
-                        &create_default_account(),
+                        sysvar::recent_blockhashes::id(),
+                        create_default_account(),
                     ),
                 ],
             ),
@@ -1558,58 +1537,57 @@ mod tests {
 
     #[test]
     fn test_process_nonce_ix_ok() {
-        let nonce_acc = nonce_account::create_account(1_000_000);
+        let nonce_account = Rc::new(nonce_account::create_account(1_000_000));
         process_instruction(
-            &Pubkey::default(),
             &serialize(&SystemInstruction::InitializeNonceAccount(Pubkey::default())).unwrap(),
             &[
-                (true, false, &Pubkey::default(), &nonce_acc),
+                (true, false, Pubkey::default(), nonce_account.clone()),
                 (
                     false,
                     false,
                     #[allow(deprecated)]
-                    &sysvar::recent_blockhashes::id(),
-                    &create_default_recent_blockhashes_account(),
+                    sysvar::recent_blockhashes::id(),
+                    create_default_recent_blockhashes_account(),
                 ),
                 (
                     false,
                     false,
-                    &sysvar::rent::id(),
-                    &create_default_rent_account(),
+                    sysvar::rent::id(),
+                    create_default_rent_account(),
                 ),
             ],
         )
         .unwrap();
-        let blockhash = &hash(&serialize(&0).unwrap());
-        let new_recent_blockhashes_account = RefCell::new(
-            #[allow(deprecated)]
+        let blockhash = hash(&serialize(&0).unwrap());
+        #[allow(deprecated)]
+        let new_recent_blockhashes_account = Rc::new(RefCell::new(
             solana_sdk::recent_blockhashes_account::create_account_with_data_for_test(
-                vec![IterItem(0u64, blockhash, 0); sysvar::recent_blockhashes::MAX_ENTRIES]
+                vec![IterItem(0u64, &blockhash, 0); sysvar::recent_blockhashes::MAX_ENTRIES]
                     .into_iter(),
             ),
-        );
-        let owner = Pubkey::default();
+        ));
         #[allow(deprecated)]
         let blockhash_id = sysvar::recent_blockhashes::id();
-        let processor_account = RefCell::new(AccountSharedData::from(Account {
-            owner: solana_sdk::native_loader::id(),
-            ..Account::default()
-        }));
         let keyed_accounts = [
-            (false, false, &Pubkey::default(), &processor_account),
-            (true, false, &owner, &nonce_acc),
-            (false, false, &blockhash_id, &new_recent_blockhashes_account),
+            (true, false, Pubkey::default(), nonce_account),
+            (false, false, blockhash_id, new_recent_blockhashes_account),
         ];
-        let mut invoke_context = &mut MockInvokeContext::new(
-            &Pubkey::default(),
-            create_keyed_accounts_unified(&keyed_accounts),
-        );
-        invoke_context.blockhash = *blockhash;
         assert_eq!(
-            super::process_instruction(
-                1,
+            mock_process_instruction(
+                &system_program::id(),
+                Vec::new(),
                 &serialize(&SystemInstruction::AdvanceNonceAccount).unwrap(),
-                invoke_context,
+                &keyed_accounts,
+                |first_instruction_account: usize,
+                 instruction_data: &[u8],
+                 invoke_context: &mut dyn InvokeContext| {
+                    invoke_context.set_blockhash(hash(&serialize(&0).unwrap()));
+                    super::process_instruction(
+                        first_instruction_account,
+                        instruction_data,
+                        invoke_context,
+                    )
+                },
             ),
             Ok(()),
         );
@@ -1632,7 +1610,6 @@ mod tests {
     fn test_process_withdraw_ix_no_keyed_accs_fail() {
         assert_eq!(
             process_instruction(
-                &Pubkey::default(),
                 &serialize(&SystemInstruction::WithdrawNonceAccount(42)).unwrap(),
                 &[],
             ),
@@ -1644,9 +1621,8 @@ mod tests {
     fn test_process_withdraw_ix_only_nonce_acc_fail() {
         assert_eq!(
             process_instruction(
-                &Pubkey::default(),
                 &serialize(&SystemInstruction::WithdrawNonceAccount(42)).unwrap(),
-                &[(true, false, &Pubkey::default(), &create_default_account())],
+                &[(true, false, Pubkey::default(), create_default_account())],
             ),
             Err(InstructionError::NotEnoughAccountKeys),
         );
@@ -1656,17 +1632,16 @@ mod tests {
     fn test_process_withdraw_ix_bad_recent_blockhash_state_fail() {
         assert_eq!(
             process_instruction(
-                &Pubkey::default(),
                 &serialize(&SystemInstruction::WithdrawNonceAccount(42)).unwrap(),
                 &[
-                    (true, false, &Pubkey::default(), &create_default_account()),
-                    (false, false, &Pubkey::default(), &create_default_account()),
+                    (true, false, Pubkey::default(), create_default_account()),
+                    (false, false, Pubkey::default(), create_default_account()),
                     (
                         false,
                         false,
                         #[allow(deprecated)]
-                        &sysvar::recent_blockhashes::id(),
-                        &create_default_account()
+                        sysvar::recent_blockhashes::id(),
+                        create_default_account()
                     ),
                 ],
             ),
@@ -1678,24 +1653,23 @@ mod tests {
     fn test_process_withdraw_ix_bad_rent_state_fail() {
         assert_eq!(
             process_instruction(
-                &Pubkey::default(),
                 &serialize(&SystemInstruction::WithdrawNonceAccount(42)).unwrap(),
                 &[
                     (
                         true,
                         false,
-                        &Pubkey::default(),
-                        &nonce_account::create_account(1_000_000),
+                        Pubkey::default(),
+                        Rc::new(nonce_account::create_account(1_000_000)),
                     ),
-                    (true, false, &Pubkey::default(), &create_default_account()),
+                    (true, false, Pubkey::default(), create_default_account()),
                     (
                         false,
                         false,
                         #[allow(deprecated)]
-                        &sysvar::recent_blockhashes::id(),
-                        &create_default_recent_blockhashes_account(),
+                        sysvar::recent_blockhashes::id(),
+                        create_default_recent_blockhashes_account(),
                     ),
-                    (false, false, &sysvar::rent::id(), &create_default_account()),
+                    (false, false, sysvar::rent::id(), create_default_account()),
                 ],
             ),
             Err(InstructionError::InvalidArgument),
@@ -1706,28 +1680,27 @@ mod tests {
     fn test_process_withdraw_ix_ok() {
         assert_eq!(
             process_instruction(
-                &Pubkey::default(),
                 &serialize(&SystemInstruction::WithdrawNonceAccount(42)).unwrap(),
                 &[
                     (
                         true,
                         false,
-                        &Pubkey::default(),
-                        &nonce_account::create_account(1_000_000),
+                        Pubkey::default(),
+                        Rc::new(nonce_account::create_account(1_000_000)),
                     ),
-                    (true, false, &Pubkey::default(), &create_default_account()),
+                    (true, false, Pubkey::default(), create_default_account()),
                     (
                         false,
                         false,
                         #[allow(deprecated)]
-                        &sysvar::recent_blockhashes::id(),
-                        &create_default_recent_blockhashes_account(),
+                        sysvar::recent_blockhashes::id(),
+                        create_default_recent_blockhashes_account(),
                     ),
                     (
                         false,
                         false,
-                        &sysvar::rent::id(),
-                        &create_default_rent_account()
+                        sysvar::rent::id(),
+                        create_default_rent_account()
                     ),
                 ],
             ),
@@ -1739,7 +1712,6 @@ mod tests {
     fn test_process_initialize_ix_no_keyed_accs_fail() {
         assert_eq!(
             process_instruction(
-                &Pubkey::default(),
                 &serialize(&SystemInstruction::InitializeNonceAccount(Pubkey::default())).unwrap(),
                 &[],
             ),
@@ -1751,13 +1723,12 @@ mod tests {
     fn test_process_initialize_ix_only_nonce_acc_fail() {
         assert_eq!(
             process_instruction(
-                &Pubkey::default(),
                 &serialize(&SystemInstruction::InitializeNonceAccount(Pubkey::default())).unwrap(),
                 &[(
                     true,
                     false,
-                    &Pubkey::default(),
-                    &nonce_account::create_account(1_000_000),
+                    Pubkey::default(),
+                    Rc::new(nonce_account::create_account(1_000_000)),
                 )],
             ),
             Err(InstructionError::NotEnoughAccountKeys),
@@ -1768,21 +1739,20 @@ mod tests {
     fn test_process_initialize_bad_recent_blockhash_state_fail() {
         assert_eq!(
             process_instruction(
-                &Pubkey::default(),
                 &serialize(&SystemInstruction::InitializeNonceAccount(Pubkey::default())).unwrap(),
                 &[
                     (
                         true,
                         false,
-                        &Pubkey::default(),
-                        &nonce_account::create_account(1_000_000),
+                        Pubkey::default(),
+                        Rc::new(nonce_account::create_account(1_000_000)),
                     ),
                     (
                         true,
                         false,
                         #[allow(deprecated)]
-                        &sysvar::recent_blockhashes::id(),
-                        &create_default_account()
+                        sysvar::recent_blockhashes::id(),
+                        create_default_account()
                     ),
                 ],
             ),
@@ -1794,23 +1764,22 @@ mod tests {
     fn test_process_initialize_ix_bad_rent_state_fail() {
         assert_eq!(
             process_instruction(
-                &Pubkey::default(),
                 &serialize(&SystemInstruction::InitializeNonceAccount(Pubkey::default())).unwrap(),
                 &[
                     (
                         true,
                         false,
-                        &Pubkey::default(),
-                        &nonce_account::create_account(1_000_000),
+                        Pubkey::default(),
+                        Rc::new(nonce_account::create_account(1_000_000)),
                     ),
                     (
                         false,
                         false,
                         #[allow(deprecated)]
-                        &sysvar::recent_blockhashes::id(),
-                        &create_default_recent_blockhashes_account(),
+                        sysvar::recent_blockhashes::id(),
+                        create_default_recent_blockhashes_account(),
                     ),
-                    (false, false, &sysvar::rent::id(), &create_default_account()),
+                    (false, false, sysvar::rent::id(), create_default_account()),
                 ],
             ),
             Err(InstructionError::InvalidArgument),
@@ -1821,27 +1790,26 @@ mod tests {
     fn test_process_initialize_ix_ok() {
         assert_eq!(
             process_instruction(
-                &Pubkey::default(),
                 &serialize(&SystemInstruction::InitializeNonceAccount(Pubkey::default())).unwrap(),
                 &[
                     (
                         true,
                         false,
-                        &Pubkey::default(),
-                        &nonce_account::create_account(1_000_000),
+                        Pubkey::default(),
+                        Rc::new(nonce_account::create_account(1_000_000)),
                     ),
                     (
                         false,
                         false,
                         #[allow(deprecated)]
-                        &sysvar::recent_blockhashes::id(),
-                        &create_default_recent_blockhashes_account(),
+                        sysvar::recent_blockhashes::id(),
+                        create_default_recent_blockhashes_account(),
                     ),
                     (
                         false,
                         false,
-                        &sysvar::rent::id(),
-                        &create_default_rent_account()
+                        sysvar::rent::id(),
+                        create_default_rent_account()
                     ),
                 ],
             ),
@@ -1851,33 +1819,31 @@ mod tests {
 
     #[test]
     fn test_process_authorize_ix_ok() {
-        let nonce_acc = nonce_account::create_account(1_000_000);
+        let nonce_account = Rc::new(nonce_account::create_account(1_000_000));
         process_instruction(
-            &Pubkey::default(),
             &serialize(&SystemInstruction::InitializeNonceAccount(Pubkey::default())).unwrap(),
             &[
-                (true, false, &Pubkey::default(), &nonce_acc),
+                (true, false, Pubkey::default(), nonce_account.clone()),
                 (
                     false,
                     false,
                     #[allow(deprecated)]
-                    &sysvar::recent_blockhashes::id(),
-                    &create_default_recent_blockhashes_account(),
+                    sysvar::recent_blockhashes::id(),
+                    create_default_recent_blockhashes_account(),
                 ),
                 (
                     false,
                     false,
-                    &sysvar::rent::id(),
-                    &create_default_rent_account(),
+                    sysvar::rent::id(),
+                    create_default_rent_account(),
                 ),
             ],
         )
         .unwrap();
         assert_eq!(
             process_instruction(
-                &Pubkey::default(),
                 &serialize(&SystemInstruction::AuthorizeNonceAccount(Pubkey::default())).unwrap(),
-                &[(true, false, &Pubkey::default(), &nonce_acc)],
+                &[(true, false, Pubkey::default(), nonce_account)],
             ),
             Ok(()),
         );
@@ -1942,7 +1908,7 @@ mod tests {
             &nonce::state::Versions::new_current(nonce::State::Initialized(
                 nonce::state::Data::default(),
             )),
-            &solana_sdk::pubkey::new_rand(),
+            &Pubkey::new_unique(),
         )
         .unwrap();
         assert_eq!(get_system_account_kind(&nonce_account), None);
@@ -1950,31 +1916,30 @@ mod tests {
 
     #[test]
     fn test_nonce_initialize_with_empty_recent_blockhashes_fail() {
-        let nonce_acc = nonce_account::create_account(1_000_000);
-        let new_recent_blockhashes_account = RefCell::new(
-            #[allow(deprecated)]
+        let nonce_account = Rc::new(nonce_account::create_account(1_000_000));
+        #[allow(deprecated)]
+        let new_recent_blockhashes_account = Rc::new(RefCell::new(
             solana_sdk::recent_blockhashes_account::create_account_with_data_for_test(
                 vec![].into_iter(),
             ),
-        );
+        ));
         assert_eq!(
             process_instruction(
-                &Pubkey::default(),
                 &serialize(&SystemInstruction::InitializeNonceAccount(Pubkey::default())).unwrap(),
                 &[
-                    (true, false, &Pubkey::default(), &nonce_acc),
+                    (true, false, Pubkey::default(), nonce_account),
                     (
                         false,
                         false,
                         #[allow(deprecated)]
-                        &sysvar::recent_blockhashes::id(),
-                        &new_recent_blockhashes_account,
+                        sysvar::recent_blockhashes::id(),
+                        new_recent_blockhashes_account,
                     ),
                     (
                         false,
                         false,
-                        &sysvar::rent::id(),
-                        &create_default_rent_account()
+                        sysvar::rent::id(),
+                        create_default_rent_account()
                     ),
                 ],
             ),
@@ -1984,57 +1949,55 @@ mod tests {
 
     #[test]
     fn test_nonce_advance_with_empty_recent_blockhashes_fail() {
-        let nonce_acc = nonce_account::create_account(1_000_000);
+        let nonce_account = Rc::new(nonce_account::create_account(1_000_000));
         process_instruction(
-            &Pubkey::default(),
             &serialize(&SystemInstruction::InitializeNonceAccount(Pubkey::default())).unwrap(),
             &[
-                (true, false, &Pubkey::default(), &nonce_acc),
+                (true, false, Pubkey::default(), nonce_account.clone()),
                 (
                     false,
                     false,
                     #[allow(deprecated)]
-                    &sysvar::recent_blockhashes::id(),
-                    &create_default_recent_blockhashes_account(),
+                    sysvar::recent_blockhashes::id(),
+                    create_default_recent_blockhashes_account(),
                 ),
                 (
                     false,
                     false,
-                    &sysvar::rent::id(),
-                    &create_default_rent_account(),
+                    sysvar::rent::id(),
+                    create_default_rent_account(),
                 ),
             ],
         )
         .unwrap();
-        let blockhash = &hash(&serialize(&0).unwrap());
-        let new_recent_blockhashes_account = RefCell::new(
-            #[allow(deprecated)]
+        #[allow(deprecated)]
+        let new_recent_blockhashes_account = Rc::new(RefCell::new(
             solana_sdk::recent_blockhashes_account::create_account_with_data_for_test(
                 vec![].into_iter(),
             ),
-        );
-        let owner = Pubkey::default();
+        ));
         #[allow(deprecated)]
         let blockhash_id = sysvar::recent_blockhashes::id();
-        let processor_account = RefCell::new(AccountSharedData::from(Account {
-            owner: solana_sdk::native_loader::id(),
-            ..Account::default()
-        }));
         let keyed_accounts = [
-            (false, false, &Pubkey::default(), &processor_account),
-            (true, false, &owner, &nonce_acc),
-            (false, false, &blockhash_id, &new_recent_blockhashes_account),
+            (true, false, Pubkey::default(), nonce_account),
+            (false, false, blockhash_id, new_recent_blockhashes_account),
         ];
-        let mut invoke_context = &mut MockInvokeContext::new(
-            &Pubkey::default(),
-            create_keyed_accounts_unified(&keyed_accounts),
-        );
-        invoke_context.blockhash = *blockhash;
         assert_eq!(
-            super::process_instruction(
-                1,
+            mock_process_instruction(
+                &system_program::id(),
+                Vec::new(),
                 &serialize(&SystemInstruction::AdvanceNonceAccount).unwrap(),
-                invoke_context,
+                &keyed_accounts,
+                |first_instruction_account: usize,
+                 instruction_data: &[u8],
+                 invoke_context: &mut dyn InvokeContext| {
+                    invoke_context.set_blockhash(hash(&serialize(&0).unwrap()));
+                    super::process_instruction(
+                        first_instruction_account,
+                        instruction_data,
+                        invoke_context,
+                    )
+                },
             ),
             Err(NonceError::NoRecentBlockhashes.into()),
         );
